Remove unnecessary includes
Avoids the linker problems that appear in the CI builds that build podio
on the fly. For some reason in these builds the presence of these
includes tricks the compiler to pull in a call to TVersionCheck which
usually happens in libpodioRootIO, even though no functionality from
there is used in these tests.