Do not build and extra library for the old schemas (#452)
* Do not build and extra library for the old schemas

Rather than building a core (current version) EDM4hep library and a
edm4hepOldSchemas library (+ corresponding dictionaries) merge
everything into the EDM4hep library. This makes it effectively
impossible to only load parts of the necessary dictionaries, e.g. to
read old files. Previously we had to make sure that we also load the old
schema library, which we only ever did properly for the python bindings,
but would have had to be done "somewhere" as well for all c++ parts.

This new approach requires the merging of two selection.xml files, for
which we now have a small python script that we run during cmake.

* Do not install no longer existing files

* Shorten the merging script

Tailor it to exactly the use case we have at the moment